Update App.tsx to enable username storage feature using localstorage
I have made changes in file frontend/src/App.tsx 

The changes made enable us to store the username in the local storage and use it when the user visits site again using the same device. Thus we need not to enter our username again and again when we visit the site and it improves the user experience.